requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
filetype==1.2.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-google-genai==2.1.3
langchain-ollama==0.3.2
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.33
ollama==0.4.8
openai==1.76.0
orjson==3.10.16
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0
